factored out fields select action from entityDetailPresenter
